Add node script for automated creation of Vue component docs
This commit adds a node script (which can be invoked via the yarn script
target "docs:vuecomponents") for automated creation of Vue component
docs which formerly had to be created manually.

For now, it has 3 sources of truth:
- the file path of the Vue components: for the component name (with and
  without Sf prefix) and the atomic type,
- the Vue component file (parsed via `vue-docgen-api` which our
  storybook plugin already uses internally): for the Props, Slots and
  Events section of the component docs,
- a (new) component info Markdown (.md) file: for all the missing
  information, currently namely the component description (prose), the
  "Most common usage scenario" section and the storybook deep-link (for
  the latter at least the component-specific part).

If a component info md file isn't present for a Vue component, a warning
is logged but the script not halted. This enables us to gradually create
such component md files.

TODO
(1) Currently missing is the extraction of CSS modifiers and SCSS
    variables. This is planned for a following commit. For now, the
    affected sections remain empty (showing the default text "None.").
(2) The generated component docs (md files) do not get inserted into the
    VuePress config files (yet). Two files have to be edited manually
    after the automated generation: _config.js_ and _enhanceApp.js_
    inside _docs/.vuepress_.