Use stylelint and prettier in shared package
The previous configuration passed all files to ESLint which couldn't
SCSS files, nor was it configured properly (babel is no dependency in
shared). Now, the ESLint config is setup correctly (with prettier, ES6
and module syntax enabled).
For that, prettier had to be added to the dependencies and called from
lintstagedrc.

In order to enable SCSS linting, stylelint is used now.