#437 Fix lozad adding new IMG tags if a multi-size src is provided
By design, lozad is supposed to be added _once_ to an element. But as we
allow to change the "lazy" attribute dynamically for our SfImages, this
triggers strange behavior: The reference to the IMG element which lozad
inserts after loading isn't fully removed from the DOM and another one
is then added to the picture tag, resulting in two images.

The fix consists of removing the spurious IMG tag when the "lazy"
attribute changes from true to false.
Also, when the attribute changes from false to true, the initialization
if lozad is retriggered.

(cherry picked from commit e697cc29e57939c186738fa6485767ad0aadaad2)